Fix #356 Use SQLite for Windows < 8.1
